requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
